+++ app/api/case_ass_rule_api/case_ass_rule_api.py
@@ -295,6 +295,9 @@ class RespAssertionRuleApi(MethodView):
         if not query_ass_resp:
             return api_result(code=400, message='返回值断言id:{}数据不存在'.format(ass_resp_id))
 
+        if query_ass_resp.creator_id != g.app_user.id:
+            return api_result(code=400, message='非管理员不能删除其他人的断言！')
+
         query_ass_resp.modifier_id = g.app_user.id
         query_ass_resp.modifier = g.app_user.username
         query_ass_resp.delete()
@@ -435,6 +438,9 @@ class FieldAssertionRuleApi(MethodView):
 
         if not query_ass_field:
             return api_result(code=400, message='字段断言id:{}数据不存在'.format(ass_field_id))
+
+        if query_ass_field.creator_id != g.app_user.id:
+            return api_result(code=400, message='非管理员不能删除其他人的断言！')
 
         query_ass_field.modifier_id = g.app_user.id
         query_ass_field.modifier = g.app_user.username
